feat(sale): add stock validation for createSale and deleteSale in sale.service.js

+++ services/sale.service.js
@@ -3,17 +3,26 @@ import ClientRepository from '../repositories/client.repository.js';
 import ProductRepository from '../repositories/product.repository.js';
 
 async function createSale(sale) {
-  let error;
+  let error = '';
   if (!(await ClientRepository.getClient(sale.client_id))) {
     error = 'O client_id informado não existe.';
   }
-  if (!(await ProductRepository.getProduct(sale.product_id))) {
+  const product = await ProductRepository.getProduct(sale.product_id);
+  if (!product) {
     error += 'O product_id informado não existe.';
   }
   if (error) {
     throw new Error(error);
   }
-  return await SaleRepository.insertSale(sale);
+
+  if (product.stock > 0) {
+    sale = await SaleRepository.insertSale(sale);
+    product.stock--;
+    await ProductRepository.updateProduct(product);
+    return sale;
+  } else {
+    throw new Error('O produto informado não possui estoque.');
+  }
 }
 
 async function getSales() {
@@ -39,7 +48,15 @@ async function updateSale(sale) {
 }
 
 async function deleteSale(id) {
-  await SaleRepository.deleteSale(id);
+  const sale = await SaleRepository.getSale(id);
+  if (sale) {
+    const product = await ProductRepository.getProduct(sale.product_id);
+    await SaleRepository.deleteSale(id);
+    product.stock++;
+    await ProductRepository.updateProduct(product);
+  } else {
+    throw new Error('O id da sale informado não existe.');
+  }
 }
 
 export default {
